# contracts/openapi/peerassist-v1.json
{"components":{"schemas":{"Action":{"enum":["organization.bootstrap","organization.read","organization.manage_policy","organization.manage_members","organization.read_audit","project.read","project.create","project.manage_settings","project.manage_members","paper.read","paper.upload","paper.add_version","review_job.read","review_job.create","review_job.cancel","review_job.retry","review_event.read","artifact.read","concern.decide","report.draft","report.finalize","retention.manage","publication.grant","legacy.read","internal.tool_execute","internal.work_claim","internal.work_complete","organization_membership.grant","organization_membership.change","organization_membership.revoke","project_membership.grant","project_membership.change","project_membership.revoke","identity.disable","identity.unlink"],"title":"Action","type":"string"},"AuditEventView":{"properties":{"action":{"title":"Action","type":"string"},"actor_id":{"format":"uuid","title":"Actor Id","type":"string"},"command_id":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Command Id"},"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"identity_id":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Identity Id"},"organization_id":{"format":"uuid","title":"Organization Id","type":"string"},"outcome":{"title":"Outcome","type":"string"},"project_id":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Project Id"},"request_id":{"title":"Request Id","type":"string"},"resource_id":{"format":"uuid","title":"Resource Id","type":"string"},"resource_type":{"title":"Resource Type","type":"string"},"safe_metadata":{"additionalProperties":true,"title":"Safe Metadata","type":"object"}},"required":["id","actor_id","organization_id","project_id","action","resource_type","resource_id","outcome","request_id","created_at","identity_id","command_id","safe_metadata"],"title":"AuditEventView","type":"object"},"CreateProjectBody":{"properties":{"name":{"maxLength":255,"minLength":1,"title":"Name","type":"string"}},"required":["name"],"title":"CreateProjectBody","type":"object"},"GrantOrganizationMembershipBody":{"properties":{"expected_version":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expected Version"},"user_id":{"format":"uuid","title":"User Id","type":"string"}},"required":["user_id"],"title":"GrantOrganizationMembershipBody","type":"object"},"GrantProjectMembershipBody":{"properties":{"expected_version":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Expected Version"},"role":{"enum":["project_owner","reviewer","viewer"],"title":"Role","type":"string"},"user_id":{"format":"uuid","title":"User Id","type":"string"}},"required":["user_id","role"],"title":"GrantProjectMembershipBody","type":"object"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"title":"Detail","type":"array"}},"title":"HTTPValidationError","type":"object"},"HealthView":{"properties":{"status":{"const":"ok","title":"Status","type":"string"}},"required":["status"],"title":"HealthView","type":"object"},"OrganizationMembershipView":{"properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"organization_id":{"format":"uuid","title":"Organization Id","type":"string"},"revoked_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Revoked At"},"role":{"title":"Role","type":"string"},"status":{"title":"Status","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"},"user_id":{"format":"uuid","title":"User Id","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","organization_id","user_id","role","status","version","created_at","updated_at","revoked_at"],"title":"OrganizationMembershipView","type":"object"},"OrganizationView":{"properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"slug":{"title":"Slug","type":"string"},"status":{"title":"Status","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","slug","name","status","version","created_at","updated_at"],"title":"OrganizationView","type":"object"},"ProjectMembershipView":{"properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"organization_id":{"format":"uuid","title":"Organization Id","type":"string"},"project_id":{"format":"uuid","title":"Project Id","type":"string"},"revoked_at":{"anyOf":[{"format":"date-time","type":"string"},{"type":"null"}],"title":"Revoked At"},"role":{"title":"Role","type":"string"},"status":{"title":"Status","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"},"user_id":{"format":"uuid","title":"User Id","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","organization_id","project_id","user_id","role","status","version","created_at","updated_at","revoked_at"],"title":"ProjectMembershipView","type":"object"},"ProjectView":{"properties":{"created_at":{"format":"date-time","title":"Created At","type":"string"},"id":{"format":"uuid","title":"Id","type":"string"},"name":{"title":"Name","type":"string"},"organization_id":{"format":"uuid","title":"Organization Id","type":"string"},"status":{"title":"Status","type":"string"},"updated_at":{"format":"date-time","title":"Updated At","type":"string"},"version":{"title":"Version","type":"integer"}},"required":["id","organization_id","name","status","version","created_at","updated_at"],"title":"ProjectView","type":"object"},"ReadinessView":{"properties":{"dependencies":{"additionalProperties":{"enum":["ready","unavailable"],"type":"string"},"title":"Dependencies","type":"object"},"status":{"enum":["ready","unavailable"],"title":"Status","type":"string"}},"required":["status","dependencies"],"title":"ReadinessView","type":"object"},"UpdateOrganizationMembershipBody":{"properties":{"expected_version":{"title":"Expected Version","type":"integer"},"status":{"enum":["active","revoked"],"title":"Status","type":"string"}},"required":["status","expected_version"],"title":"UpdateOrganizationMembershipBody","type":"object"},"UpdateProjectMembershipBody":{"properties":{"expected_version":{"title":"Expected Version","type":"integer"},"role":{"enum":["project_owner","reviewer","viewer"],"title":"Role","type":"string"},"status":{"enum":["active","revoked"],"title":"Status","type":"string"}},"required":["role","status","expected_version"],"title":"UpdateProjectMembershipBody","type":"object"},"ValidationError":{"properties":{"ctx":{"title":"Context","type":"object"},"input":{"title":"Input"},"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"title":"Location","type":"array"},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}},"required":["loc","msg","type"],"title":"ValidationError","type":"object"}}},"info":{"title":"PeerAssist Platform API","version":"1.0.0"},"openapi":"3.1.0","paths":{"/api/v1/auth/callback":{"get":{"operationId":"v1_auth_callback","parameters":[{"in":"query","name":"state","required":true,"schema":{"title":"State","type":"string"}},{"in":"query","name":"code","required":true,"schema":{"title":"Code","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Callback","tags":["authentication"]}},"/api/v1/auth/login":{"get":{"operationId":"v1_auth_login","parameters":[{"in":"query","name":"return_path","required":false,"schema":{"default":"/","title":"Return Path","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Login","tags":["authentication"]}},"/api/v1/auth/logout":{"post":{"operationId":"v1_auth_logout","parameters":[{"in":"header","name":"X-CSRF-Token","required":true,"schema":{"maxLength":512,"minLength":1,"title":"X-Csrf-Token","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Logout","tags":["authentication"]}},"/api/v1/auth/session":{"get":{"operationId":"v1_auth_session","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response V1 Auth Session","type":"object"}}},"description":"Successful Response"}},"summary":"Session","tags":["authentication"]}},"/api/v1/health":{"get":{"operationId":"v1_health","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HealthView"}}},"description":"Successful Response"}},"summary":"Health","tags":["system"]}},"/api/v1/me":{"get":{"operationId":"v1_me","responses":{"200":{"content":{"application/json":{"schema":{"additionalProperties":true,"title":"Response V1 Me","type":"object"}}},"description":"Successful Response"}},"summary":"Me","tags":["authentication"]}},"/api/v1/organizations":{"get":{"operationId":"v1_list_organizations","responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrganizationView"},"title":"Response V1 List Organizations","type":"array"}}},"description":"Successful Response"}},"summary":"List Organizations","tags":["organizations"]}},"/api/v1/organizations/{organization_id}/audit-events":{"get":{"operationId":"v1_list_organization_audit_events","parameters":[{"in":"path","name":"organization_id","required":true,"schema":{"format":"uuid","title":"Organization Id","type":"string"}},{"in":"query","name":"project_id","required":false,"schema":{"anyOf":[{"format":"uuid","type":"string"},{"type":"null"}],"title":"Project Id"}},{"in":"query","name":"action","required":false,"schema":{"anyOf":[{"$ref":"#/components/schemas/Action"},{"type":"null"}],"title":"Action"}},{"in":"query","name":"outcome","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Outcome"}},{"in":"query","name":"resource_type","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Resource Type"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/AuditEventView"},"title":"Response V1 List Organization Audit Events","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Audit Events","tags":["organizations"]}},"/api/v1/organizations/{organization_id}/members":{"get":{"operationId":"v1_list_organization_members","parameters":[{"in":"path","name":"organization_id","required":true,"schema":{"format":"uuid","title":"Organization Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/OrganizationMembershipView"},"title":"Response V1 List Organization Members","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Members","tags":["organizations"]},"post":{"operationId":"v1_grant_organization_member","parameters":[{"in":"path","name":"organization_id","required":true,"schema":{"format":"uuid","title":"Organization Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"maxLength":255,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrantOrganizationMembershipBody"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationMembershipView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Grant Member","tags":["organizations"]}},"/api/v1/organizations/{organization_id}/members/{membership_id}":{"patch":{"operationId":"v1_update_organization_member","parameters":[{"in":"path","name":"organization_id","required":true,"schema":{"format":"uuid","title":"Organization Id","type":"string"}},{"in":"path","name":"membership_id","required":true,"schema":{"format":"uuid","title":"Membership Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"maxLength":255,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateOrganizationMembershipBody"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/OrganizationMembershipView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Update Member","tags":["organizations"]}},"/api/v1/organizations/{organization_id}/projects":{"get":{"operationId":"v1_list_organization_projects","parameters":[{"in":"path","name":"organization_id","required":true,"schema":{"format":"uuid","title":"Organization Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectView"},"title":"Response V1 List Organization Projects","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Projects","tags":["organizations"]},"post":{"operationId":"v1_create_project","parameters":[{"in":"path","name":"organization_id","required":true,"schema":{"format":"uuid","title":"Organization Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"maxLength":255,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateProjectBody"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Create Project","tags":["organizations"]}},"/api/v1/projects/{project_id}":{"get":{"operationId":"v1_get_project","parameters":[{"in":"path","name":"project_id","required":true,"schema":{"format":"uuid","title":"Project Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Get Project","tags":["projects"]}},"/api/v1/projects/{project_id}/members":{"get":{"operationId":"v1_list_project_members","parameters":[{"in":"path","name":"project_id","required":true,"schema":{"format":"uuid","title":"Project Id","type":"string"}}],"responses":{"200":{"content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProjectMembershipView"},"title":"Response V1 List Project Members","type":"array"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"List Members","tags":["projects"]},"post":{"operationId":"v1_grant_project_member","parameters":[{"in":"path","name":"project_id","required":true,"schema":{"format":"uuid","title":"Project Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"maxLength":255,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GrantProjectMembershipBody"}}},"required":true},"responses":{"201":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMembershipView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Grant Member","tags":["projects"]}},"/api/v1/projects/{project_id}/members/{membership_id}":{"patch":{"operationId":"v1_update_project_member","parameters":[{"in":"path","name":"project_id","required":true,"schema":{"format":"uuid","title":"Project Id","type":"string"}},{"in":"path","name":"membership_id","required":true,"schema":{"format":"uuid","title":"Membership Id","type":"string"}},{"in":"header","name":"Idempotency-Key","required":true,"schema":{"maxLength":255,"minLength":1,"title":"Idempotency-Key","type":"string"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateProjectMembershipBody"}}},"required":true},"responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProjectMembershipView"}}},"description":"Successful Response"},"422":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}},"description":"Validation Error"}},"summary":"Update Member","tags":["projects"]}},"/api/v1/ready":{"get":{"operationId":"v1_ready","responses":{"200":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadinessView"}}},"description":"Successful Response"},"503":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ReadinessView"}}},"description":"Service Unavailable"}},"summary":"Readiness","tags":["system"]}}}}